requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.10
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
whitenoise==5.1.0
